pyproject.toml
[tool.poetry]
name = "animaldonor"
version = "0.1.0"
description = ""
authors = ["PrinterOMG <PrinterOMG@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
sqlalchemy = {version = "^2.0.27", extras = ["asyncio"]}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.6.1"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
